Rewrite xdp_document_for_uri to be async
The sync version of this command did only look in the
documents hash table, which does not work, because
documents are only loaded from the database on demand,
so the hash table is initially empty, and we happily
insert duplicates into the database.

The new, async version looks up an existing document
by uri before creating a new one. The async version
is more code, but we can at least share the async
permission loading code with the xdp_document_load
implementation.