Fix up xdp_document_for_uri
We were returning the document object before the async saving
to the database was complete. At that point, the id of the
document is still 0, and our following grant_permissions call
was storing permissions for this nonexisting document with
id 0. Fix this by only returning the document after it has
been saved.