Filter attributes in GetInfo
Filter attributes we allow to avoid leaking any detailed data from the system
that is not necessarily part of the document per-se, such as exact modification
times, filesystem or device ids or xattrs.

Anything that you can get from just reading the file is ok, like file size,
mime type and thumbnail info.

There are also some things that are implied with the user allowing access to the
file, such as file name and icons.

A few less obvious things are needed to a good UI. This includes the etag, and
access rights.