Drop window arguments
It seems better to keep the document portal a UI-less
service that just handles handle<>fd translation. If
errors need to be reported to the user, we can use
the app id to find name, icon, etc to represent the
app at fault.

We don't have window ids for Wayland surfaces, anyway.