Add AddLocal portal method
This allows a contained app to still create a document, providing it
can prove that it has access to the file. It does so by opening the
file and passing a fd to the portal, which verifies that the file is
the same outside the container before creating the document and
granting it permissions (based on the fd open mode).