Simplify deletion
Instead of queueing up deletions of individual permission objects,
just assume that the db removal will always work, and return right
away. This changes the behaviour for parallel revocations so that
only the first one will succeed, and all later RevokePermissions
calls for the same handle will return an error.