Actually report delete errors
During deletion we keep the document in the hash, but mark it as
deleting. Then we fully remove it when the operation succeeds, or
marks it as non-deleting if that fails. This lets us properly report
the failure to the initial deleter, so we don't accidentally get
leftover documents.

If we just remove it from the hash immediately, then we will look it
up from the db and restore it if someone accesses it. Now instead we
know from the existing document that we're not supposed to do that,
and we treat it as if it wasn't there.

We also make sure a different Add for the same uri does not reuse the
deleting document.

Semantically its a bit weird that the document is ressurected after
having been "deleted" for a while, but this should not be common and
is the best we can do imho.