Add fuse implementation for document access
This mounts a virtual filesystem at /run/user/$uid/xdp which contains
all the documents as a directory, with only the one file
inside. Writing to this file only ever does atomic replace, and once
a file has been written all the open fds to the written files are
made readonly to guarantee that the file is not further changed.

There is also a by-app subdirectory with a further directory in where
only the documents available to that app is visible.